Minimal presto-pinot driver exposing just pinot-common and some helper
This is so that the accompanying presto-pinot connector in
prestodb/presto proper has simpler dependencies.

This is in line with the approach taken by other heavy weight connectors
like presto-kudu and presto-cassandra.

The basic approach is that the stuff in pinot-common is exposed out as
is and the rest are shaded, since this is the public api. Also added a
class taken from the Pinot broker to federate a PQL to bunch of servers.

The helper class PinotScatterGatherQueryClient does the job of the pinot
to send the query to a pinot server via the internal thrift protocol.

TODO: Couldn't figure out what to do with the logging classes. Currently
they aren't exposed out but what happens if this library need to do some
logging ?